+++ src/modules/custom-plugins/plugins-settings-ui/plugins-settings-ui.module.ts
@@ -5,7 +5,6 @@ import { AuthModule } from '../../../core/auth/auth.module.js'
 import { ConfigModule } from '../../../core/config/config.module.js'
 import { LoggerModule } from '../../../core/logger/logger.module.js'
 import { PluginsModule } from '../../plugins/plugins.module.js'
-import { PluginsSettingsUiTicketService } from './plugins-settings-ui-ticket.service.js'
 import { PluginsSettingsUiController } from './plugins-settings-ui.controller.js'
 import { PluginsSettingsUiGateway } from './plugins-settings-ui.gateway.js'
 import { PluginsSettingsUiService } from './plugins-settings-ui.service.js'
@@ -20,7 +19,6 @@ import { PluginsSettingsUiService } from './plugins-settings-ui.service.js'
   ],
   providers: [
     PluginsSettingsUiService,
-    PluginsSettingsUiTicketService,
     PluginsSettingsUiGateway,
   ],
   controllers: [